refactor and repair Attacks
