+++ src/main/java/com/java_academy/logic/attacks/NormalAttack.java
@@ -1,9 +1,10 @@
-package com.java_academy.logic.AttackLogic;
+package com.java_academy.logic.attacks;
 
 /**
  * Created by Bartłomiej Janik on 7/26/2017.
  */
 public class NormalAttack implements Attack {
+	
     @Override
     public void attack(int position) {
         System.out.println(position);
